// src/lib/api/generated/models/v1ExternalOTCSide.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * ExternalOTCSide identifies which side a thread iteration came
 * from, scoped to this bank.
 */
export enum v1ExternalOTCSide {
    EXTERNAL_OTC_SIDE_UNSPECIFIED = 'EXTERNAL_OTC_SIDE_UNSPECIFIED',
    EXTERNAL_OTC_SIDE_LOCAL = 'EXTERNAL_OTC_SIDE_LOCAL',
    EXTERNAL_OTC_SIDE_REMOTE = 'EXTERNAL_OTC_SIDE_REMOTE',
}