// src/lib/api/generated/models/v1ExternalOTCRole.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * ExternalOTCRole says whether the local party is buyer or seller.
 */
export enum v1ExternalOTCRole {
    EXTERNAL_OTC_ROLE_UNSPECIFIED = 'EXTERNAL_OTC_ROLE_UNSPECIFIED',
    EXTERNAL_OTC_ROLE_BUYER = 'EXTERNAL_OTC_ROLE_BUYER',
    EXTERNAL_OTC_ROLE_SELLER = 'EXTERNAL_OTC_ROLE_SELLER',
}