// src/lib/api/generated/models/v1AcceptExternalOTCOfferResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { v1ExternalOTCContract } from './v1ExternalOTCContract';
import type { v1ExternalOTCThread } from './v1ExternalOTCThread';
export type v1AcceptExternalOTCOfferResponse = {
    thread?: v1ExternalOTCThread;
    contract?: v1ExternalOTCContract;
};